pyproject.toml
[tool.poetry]
name = "2023-2-squad06"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2 = "^2.9.7"
django-environ = "^0.11.2"
django-guardian = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.9.1"
isort = "^5.12.0"
commitizen = "^3.9.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.4"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.1.2"
sphinxawesome-theme = {version = "^5.0.0b3", allow-prereleases = true}
sphinx-design = "^0.5.0"

[tool.black]
color = true
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
